combine function that generate mesh in the script that create the matter profile file for the fermi-dirac test
